refactor(routines): F4.2.0 routine.rs + RoutineError verbatim 搬迁到 dasclaw_routines
ADR-152 §3 F4.2 / §11.8.9.3、ADR-129 §1.3。

物理搬迁（git rename 99%+）：
- desktop-client/ironclaw/src/routines/routine.rs → crates/dasclaw_routines/src/routine.rs
- RoutineError enum (15 变体) → crates/dasclaw_routines/src/error.rs

import 重写（verbatim 仅做必要绑定调整）：
- crate::timezone::parse_timezone → dasclaw_workspace_cap::timezone::parse_timezone (2 处)
- 测试 use crate::routines::routine::{...} → use crate::routine::{...}
- MAX_TOOL_ROUNDS_LIMIT pub(crate) → pub（跨 crate 后唯一调用方仍需访问）

desktop 侧兼容 re-export（不动调用点）：
- error.rs: pub use dasclaw_routines::RoutineError;
- routines/mod.rs: pub use dasclaw_routines::routine;

非目标：routine_engine / scheduler / cost_guard / heartbeat / job_monitor /
self_repair 因仍密集反向依赖 14 个 desktop 模块，留待后续切片
（详见 docs/plans/architecture-refactor/f42-preflight-dependency-assessment.md）。

验证：
- cargo check -p dasclaw_routines --tests ✅
- cargo check -p dasclaw --tests ✅
- cargo nextest run -p dasclaw_routines ✅ 36/36
- cargo clippy --no-deps -p dasclaw_routines --all-targets -- -D warnings ✅
- cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings ✅
- scripts/check_no_panics.py --base origin/xClaw ✅
- cargo fmt --all ✅

Closes #727
