+++ crates/dasclaw_routines/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "dasclaw_routines"
 version = "0.0.0-w1"
-edition = "2021"
+edition = "2024"
 description = "Routine orchestration (fork ironclaw 0.24 routines module promotion)."
 license = "Apache-2.0 OR MIT"
 publish = false
@@ -10,4 +10,14 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+cron = "0.13"
+hex = "0.4.3"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.10"
 thiserror = "1"
+tracing = "0.1"
+uuid = { version = "1", features = ["v4", "serde"] }
+
+dasclaw_workspace_cap = { path = "../dasclaw_workspace_cap", version = "0.1.0" }
